server/requirements.txt
pycryptodome
Pyro5
flask
flask_sse
apscheduler
flask_cors
sse-starlette
APScheduler